Reworking the logging services, using Raygun.io
