+++ main/src/core/MonoDevelop.Core/MonoDevelop.Core.Assemblies/TargetRuntime.cs
@@ -388,7 +388,7 @@ namespace MonoDevelop.Core.Assemblies
 				try {
 					RunInitialization ();
 				} catch (Exception ex) {
-					LogReporting.LogReportingService.ReportUnhandledException (ex, false);
+					//LogReporting.LogReportingService.ReportUnhandledException (ex, false);
 					LoggingService.LogFatalError ("Unhandled exception in SystemAssemblyService background initialisation thread.", ex);
 				} finally {
 					lock (initEventLock) {
