+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.CodeTemplates/CodeTemplateService.cs
@@ -155,7 +155,7 @@ namespace MonoDevelop.Ide.CodeTemplates
 			}
 			foreach (CodeTemplate template in templates) {
 				if (string.IsNullOrEmpty (template.Shortcut)) {
-					LoggingService.LogError ("CodeTemplateService: Can't save unnamed template " + template);
+					LoggingService.LogUserError ("CodeTemplateService: Can't save unnamed template " + template);
 					continue;
 				}
 				SaveTemplate (template, Path.Combine (TemplatePath, template.Shortcut + ".template.xml"));
