+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/IdeStartup.cs
@@ -123,7 +123,7 @@ namespace MonoDevelop.Ide
 			EndPoint ep = null;
 			
 			DispatchService.Initialize ();
-			
+
 			// Set a synchronization context for the main gtk thread
 			SynchronizationContext.SetSynchronizationContext (new GtkSynchronizationContext ());
 			
@@ -505,7 +505,7 @@ namespace MonoDevelop.Ide
 			LoggingService.LogError (string.Format ("An unhandled exception has occured. Terminating MonoDevelop? {0}", willShutdown), ex);
 			
 			// Pass it off to the reporting service now.
-			LogReportingService.ReportUnhandledException (ex, willShutdown);
+			LoggingService.ReportUnhandledException (ex, willShutdown);
 		}
 		
 		/// <summary>SDBM-style hash, bounded to a range of 1000.</summary>
