+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/LogReportingStartup.cs
@@ -43,11 +43,11 @@ namespace MonoDevelop.Ide
 		{
 			System.Threading.ThreadPool.QueueUserWorkItem (delegate {
 				// Process cached crash reports if there are any and uploading is enabled
-				LogReportingService.ProcessCache ();
+				LoggingService.ProcessCache ();
 			});
 			
 			// Attach a handler for when exceptions need to be processed
-			LogReportingService.UnhandledErrorOccured = (enabled, ex, willShutdown) => {
+			LoggingService.UnhandledErrorOccured = (enabled, ex, willShutdown) => {
 				var doNotSend = new AlertButton (GettextCatalog.GetString ("Do _Not Send"));
 				var sendOnce = new AlertButton (GettextCatalog.GetString ("_Send This Time"));
 				var alwaysSend = new AlertButton (GettextCatalog.GetString ("_Always Send"));
